Написал тесты и RegisterImpl
