+++ sandbox.db/src_resources/liquibase/postgres/v001/002_client.xml
@@ -6,7 +6,7 @@
   <changeSet author="anazar" id="002_client-005">
     <sql endDelimiter=";;"><![CDATA[
       create table charm (
-      id int not null,
+      id serial not null,
       name varchar(255),
       description varchar(255),
       energy real,
@@ -19,7 +19,7 @@
   <changeSet author="anazar" id="002_client-001">
     <sql endDelimiter=";;"><![CDATA[
       create table client (
-        id int not null,
+        id serial not null,
         firstname varchar(50) not null,
         lastname varchar(255) not null,
         patronymic varchar(255),
@@ -59,7 +59,7 @@
   <changeSet author="anazar" id="002_client-004">
     <sql endDelimiter=";;"><![CDATA[
       create table client_account (
-        id int not null primary key,
+        id serial not null primary key,
         client int,
         money real,
         number varchar(255),
@@ -71,7 +71,7 @@
   <changeSet author="anazar" id="002_client-007">
     <sql endDelimiter=";;"><![CDATA[
       create table transaction_type (
-      id int not null,
+      id serial not null,
       code varchar(255),
       name varchar(255),
       primary key(id)
@@ -82,7 +82,7 @@
   <changeSet author="anazar" id="002_client-006">
     <sql endDelimiter=";;"><![CDATA[
       create table client_account_transaction (
-      id int not null,
+      id serial not null,
       account int,
       money real,
       finished_at timestamp,
