--- sandbox.db/test_src/kz/greetgo/sandbox/db/test/dao/postgres/ClientTestDao1Postgres.java
@@ -1,8 +1,0 @@
-package kz.greetgo.sandbox.db.test.dao.postgres;
-
-import kz.greetgo.depinject.core.Bean;
-import kz.greetgo.sandbox.db.test.dao.ClientTestDao1;
-
-@Bean
-public interface ClientTestDao1Postgres extends ClientTestDao1 {
-}
